+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       ...
   ```
 
+- Added `PinchAudioService` with procide access to Pinch Translation.
+
 ### Deprecated
 
 - `PipelineTask` events `on_pipeline_stopped`, `on_pipeline_ended` and